+++ src/test/groovy/de/lisaplus/atlas/codegen/test/PlantUml.groovy
@@ -2,7 +2,9 @@ package de.lisaplus.atlas.codegen.test
 
 import org.junit.Test
 
+import static junit.framework.Assert.assertEquals
 import static junit.framework.Assert.assertTrue
+import static org.junit.Assert.assertFalse
 
 /**
  * Tests the plantuml generator and template
@@ -33,6 +35,106 @@ class PlantUml {
         doCodeGen.generator_parameters.add('removeEmptyLines=true')
         doCodeGen.run()
         assertTrue(new File(destFile).exists())
+    }
+
+    @Test
+    void createIncidentModelAddTags() {
+        def destFile = 'tmp/incident.puml'
+        de.lisaplus.atlas.DoCodeGen doCodeGen = new de.lisaplus.atlas.DoCodeGen()
+        doCodeGen.model = 'src/test/resources/test_schemas/ds/incident.json'
+        doCodeGen.generators.add('singlefile=src/main/resources/templates/meta/plantuml.txt')
+        doCodeGen.outputBaseDir = 'tmp'
+        doCodeGen.generator_parameters.add('destFileName=incident.puml')
+        doCodeGen.generator_parameters.add('removeEmptyLines=true')
+        doCodeGen.typeAddTagList.add('Incident=main')
+        doCodeGen.typeAddTagList.add('Incident=cool')
+        doCodeGen.typeAddTagList.add('Domain=cool')
+        doCodeGen.run()
+        assertTrue(new File(destFile).exists())
+        boolean incidentFound=false
+        boolean domainFound=false
+        doCodeGen.dataModel.types.each { type ->
+            if (type.name=='Incident') {
+                incidentFound=true
+                assertEquals(5,type.tags.size())
+                assertTrue(type.tags.contains('main'))
+                assertTrue(type.tags.contains('cool'))
+            }
+            else if (type.name=='Domain') {
+                domainFound=true
+                assertEquals(1,type.tags.size())
+                assertTrue(type.tags.contains('cool'))
+            }
+            else {
+                assertFalse(type.tags.contains('main'))
+                assertFalse(type.tags.contains('cool'))
+            }
+        }
+        assertTrue(incidentFound)
+        assertTrue(domainFound)
+    }
+
+    @Test
+    void createIncidentModelRemoveTags() {
+        def destFile = 'tmp/incident.puml'
+        de.lisaplus.atlas.DoCodeGen doCodeGen = new de.lisaplus.atlas.DoCodeGen()
+        doCodeGen.model = 'src/test/resources/test_schemas/ds/incident.json'
+        doCodeGen.generators.add('singlefile=src/main/resources/templates/meta/plantuml.txt')
+        doCodeGen.outputBaseDir = 'tmp'
+        doCodeGen.generator_parameters.add('destFileName=incident.puml')
+        doCodeGen.generator_parameters.add('removeEmptyLines=true')
+        doCodeGen.typeRemoveTagList.add('Incident=mongodb')
+        doCodeGen.typeRemoveTagList.add('Incident=rest')
+        doCodeGen.typeRemoveTagList.add('Incident=joined')
+        doCodeGen.typeRemoveTagList.add('IncidentTag=rest')
+        doCodeGen.run()
+        assertTrue(new File(destFile).exists())
+        boolean incidentFound=false
+        boolean incidentTagFound=false
+        doCodeGen.dataModel.types.each { type ->
+            if (type.name=='Incident') {
+                incidentFound=true
+                assertEquals(0,type.tags.size())
+            }
+            else if (type.name=='IncidentTag') {
+                incidentTagFound=true
+                assertEquals(1,type.tags.size())
+                assertTrue(type.tags.contains('mongodb'))
+            }
+        }
+        assertTrue(incidentFound)
+        assertTrue(incidentTagFound)
+    }
+
+    @Test
+    void createIncidentModelRemoveAllTags() {
+        def destFile = 'tmp/incident.puml'
+        de.lisaplus.atlas.DoCodeGen doCodeGen = new de.lisaplus.atlas.DoCodeGen()
+        doCodeGen.model = 'src/test/resources/test_schemas/ds/incident.json'
+        doCodeGen.generators.add('singlefile=src/main/resources/templates/meta/plantuml.txt')
+        doCodeGen.outputBaseDir = 'tmp'
+        doCodeGen.generator_parameters.add('destFileName=incident.puml')
+        doCodeGen.generator_parameters.add('removeEmptyLines=true')
+        doCodeGen.generator_parameters.add('guidTypeColor=f9f6e0')
+        doCodeGen.generator_parameters.add('ignoreUnRefTypes=true')
+        doCodeGen.typeRemoveTagAllList.add('rest')
+        doCodeGen.run()
+        assertTrue(new File(destFile).exists())
+        boolean incidentFound=false
+        boolean incidentTagFound=false
+        doCodeGen.dataModel.types.each { type ->
+            assertFalse(type.tags.contains('rest'))
+            if (type.name=='Incident') {
+                incidentFound=true
+                assertEquals(2,type.tags.size())
+            }
+            else if (type.name=='IncidentTag') {
+                incidentTagFound=true
+                assertEquals(1,type.tags.size())
+            }
+        }
+        assertTrue(incidentFound)
+        assertTrue(incidentTagFound)
     }
 
     @Test